Split objects.py into one file pr object-type
objects.py was getting relatively large and would continue to grow in
near future due to the introduction of more object types. Hence, it
make sense to split it up into multiple files. One pr object class.
